Made restoring of application window more robust
When the application window is restored by left clicking on the notification area icon, it will now reliably be given focus and brought to the foreground. Likewise, when the application window is already open, if the user left clicks the notification area icon the application window will get focus and be brought to the foreground.